refactor(content_ui): localize translation graphql
