+++ frontend/plugins/content_ui/src/modules/cms/graphql/queries.ts
@@ -797,48 +797,3 @@ export const CMS_CUSTOM_POST_TYPE_REMOVE = gql`
     cmsCustomPostTypesRemove(_id: $_id)
   }
 `;
-
-export const CMS_TRANSLATIONS = gql`
-  query cmsTranslations($objectId: String, $type: String) {
-    cmsTranslations(objectId: $objectId, type: $type) {
-      _id
-      objectId
-      language
-      title
-      content
-      excerpt
-      customFieldsData
-      type
-    }
-  }
-`;
-
-export const CMS_ADD_TRANSLATION = gql`
-  mutation cmsAddTranslation($input: TranslationInput!) {
-    cmsAddTranslation(input: $input) {
-      _id
-      objectId
-      language
-      title
-      content
-      excerpt
-      customFieldsData
-      type
-    }
-  }
-`;
-
-export const CMS_EDIT_TRANSLATION = gql`
-  mutation cmsEditTranslation($input: TranslationInput!) {
-    cmsEditTranslation(input: $input) {
-      _id
-      objectId
-      language
-      title
-      content
-      excerpt
-      customFieldsData
-      type
-    }
-  }
-`;
