+++ frontend/plugins/content_ui/src/modules/cms/shared/graphql/index.ts
@@ -1,0 +1,2 @@
+export * from './mutations';
+export * from './queries';
