+++ frontend/plugins/content_ui/src/modules/cms/shared/graphql/mutations/index.ts
@@ -1,0 +1,4 @@
+export {
+  CMS_ADD_TRANSLATION,
+  CMS_EDIT_TRANSLATION,
+} from './translationsMutations';
