+++ frontend/plugins/content_ui/src/modules/cms/shared/graphql/queries/index.ts
@@ -1,0 +1,1 @@
+export { CMS_TRANSLATIONS } from './translationsQuery';
