+++ frontend/plugins/content_ui/src/modules/cms/shared/graphql/queries/translationsQuery.ts
@@ -1,0 +1,16 @@
+import { gql } from '@apollo/client';
+
+export const CMS_TRANSLATIONS = gql`
+  query cmsTranslations($objectId: String, $type: String) {
+    cmsTranslations(objectId: $objectId, type: $type) {
+      _id
+      objectId
+      language
+      title
+      content
+      excerpt
+      customFieldsData
+      type
+    }
+  }
+`;
